requirements.txt
beautifulsoup4==4.7.1
celery==4.3.0
dj-database-url==0.5.0
Django==2.1.7
django-celery-beat==1.4.0
django-filter==2.1.0
django-rest-swagger==2.1.0
django-webpack-loader==0.6.0
djangorestframework==3.9.1
djangorestframework-gis==0.14
lxml==4.3.2
psycopg2==2.8.2
python-decouple==3.1
requests==2.21.0
